second

+++ homework1_step1.c
@@ -13,7 +13,7 @@ void entrance()
     //void* osas = main;
     //doNotTouch[10] = (long long)osas + 33;
     //void *variable = secret;
-    doNotTouch[4] = secret;
+    doNotTouch[4] = (long long)secret;
 
     // can only modify this section END
     printf("now inside entrance()!\n");
